sample: Fix warnings from Clang

+++ sample/09-softserial/softserial.h
@@ -2,7 +2,7 @@
 // softserial.h
 //
 // Circle - A C++ bare metal environment for Raspberry Pi
-// Copyright (C) 2014-2018  R. Stange <rsta2@o2online.de>
+// Copyright (C) 2014-2022  R. Stange <rsta2@o2online.de>
 // 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -35,8 +35,8 @@ public:
 
 	boolean Initialize (void);
 
-	int Write (const void *pBuffer, unsigned nCount);
-	int Read (void *pBuffer, unsigned nCount);
+	int Write (const void *pBuffer, size_t nCount);
+	int Read (void *pBuffer, size_t nCount);
 
 private:
 	void InterruptHandler (void);
